Refactored bot commands and added search functionality
With this commit, we have updated the bot commands for better functionality and included search functionality in the bot. In bot commands, start, help, setting and mine are implemented with private chat restriction. The command feature has been moved to another file for clarity. We also removed the unnecessary 'cancel' command. Simultaneously added a search module to handle search requests and produce search results, sorting requested entries by relevance. We have updated the database ChatType field modification to make it compatible with the search functionality. The worker file got updated to include methods that are globally required, along with db/types file for satisfying the search functionality. The handler implementation got modified to handle search contexts along with other context formations. The callback router was redesigned to bring clarity and modularity, it also updated the context type for better handling of search result callback. Callback messages got updated to handle the search results and show them to the user.